Fix x86 (32-bit) compatibility issues by converting Int to Int64
On x86 systems, Int is Int32 but Sundials C API expects Int64 (sunindextype).
This fixes MethodError: no method matching N_VMake_Serial(::Int32, ...).

Fixed in:
- src/nvector_wrapper.jl: Convert length(v) to Int64 in N_VMake_Serial call
- src/common_interface/solve.jl: Convert all length() calls to Int64 for SUNDenseMatrix, SUNBandMatrix, and SUNSparseMatrix
- src/simple.jl: Convert all length() calls to Int64 for matrix constructors

+++ src/common_interface/solve.jl
@@ -256,7 +256,7 @@ function DiffEqBase.__init(prob::DiffEqBase.AbstractODEProblem{uType, tupType, i
     if Method == :Newton # Only use a linear solver if it's a Newton-based method
         if LinearSolver in (:Dense, :LapackDense)
             nojacobian = false
-            A = SUNDenseMatrix(length(uvec), length(uvec), ctx)
+            A = SUNDenseMatrix(Int64(length(uvec)), Int64(length(uvec)), ctx)
             _A = MatrixHandle(A, DenseMatrix())
             if LinearSolver === :Dense
                 LS = SUNLinSol_Dense(utmp, A, ctx)
@@ -267,7 +267,7 @@ function DiffEqBase.__init(prob::DiffEqBase.AbstractODEProblem{uType, tupType, i
             end
         elseif LinearSolver in (:Band, :LapackBand)
             nojacobian = false
-            A = SUNBandMatrix(length(uvec), alg.jac_upper, alg.jac_lower, ctx)
+            A = SUNBandMatrix(Int64(length(uvec)), Int64(alg.jac_upper), Int64(alg.jac_lower), ctx)
             _A = MatrixHandle(A, BandMatrix())
             if LinearSolver === :Band
                 LS = SUNLinSol_Band(utmp, A, ctx)
@@ -304,7 +304,8 @@ function DiffEqBase.__init(prob::DiffEqBase.AbstractODEProblem{uType, tupType, i
         elseif LinearSolver == :KLU
             nojacobian = false
             nnz = length(SparseArrays.nonzeros(prob.f.jac_prototype))
-            A = SUNSparseMatrix(length(uvec), length(uvec), nnz, CSC_MAT, ctx)
+            A = SUNSparseMatrix(
+                Int64(length(uvec)), Int64(length(uvec)), Int64(nnz), CSC_MAT, ctx)
             LS = SUNLinSol_KLU(utmp, A, ctx)
             _A = MatrixHandle(A, SparseMatrix())
             _LS = LinSolHandle(LS, KLU())
@@ -731,7 +732,7 @@ function DiffEqBase.__init(prob::DiffEqBase.AbstractODEProblem{uType, tupType, i
     if Method == :Newton && alg.stiffness !== Explicit() # Only use a linear solver if it's a Newton-based method
         if LinearSolver in (:Dense, :LapackDense)
             nojacobian = false
-            A = SUNDenseMatrix(length(uvec), length(uvec), ctx)
+            A = SUNDenseMatrix(Int64(length(uvec)), Int64(length(uvec)), ctx)
             _A = MatrixHandle(A, DenseMatrix())
             if LinearSolver === :Dense
                 LS = SUNLinSol_Dense(utmp, A, ctx)
@@ -742,7 +743,7 @@ function DiffEqBase.__init(prob::DiffEqBase.AbstractODEProblem{uType, tupType, i
             end
         elseif LinearSolver in (:Band, :LapackBand)
             nojacobian = false
-            A = SUNBandMatrix(length(uvec), alg.jac_upper, alg.jac_lower, ctx)
+            A = SUNBandMatrix(Int64(length(uvec)), Int64(alg.jac_upper), Int64(alg.jac_lower), ctx)
             _A = MatrixHandle(A, BandMatrix())
             if LinearSolver === :Band
                 LS = SUNLinSol_Band(utmp, A, ctx)
@@ -773,7 +774,8 @@ function DiffEqBase.__init(prob::DiffEqBase.AbstractODEProblem{uType, tupType, i
             _LS = LinSolHandle(LS, PTFQMR())
         elseif LinearSolver == :KLU
             nnz = length(SparseArrays.nonzeros(prob.f.jac_prototype))
-            A = SUNSparseMatrix(length(uvec), length(uvec), nnz, CSC_MAT, ctx)
+            A = SUNSparseMatrix(
+                Int64(length(uvec)), Int64(length(uvec)), Int64(nnz), CSC_MAT, ctx)
             LS = SUNLinSol_KLU(utmp, A, ctx)
             _A = MatrixHandle(A, SparseMatrix())
             _LS = LinSolHandle(LS, KLU())
@@ -804,7 +806,7 @@ function DiffEqBase.__init(prob::DiffEqBase.AbstractODEProblem{uType, tupType, i
     if prob.f.mass_matrix != LinearAlgebra.I && alg.stiffness !== Explicit()
         if MassLinearSolver in (:Dense, :LapackDense)
             nojacobian = false
-            M = SUNDenseMatrix(length(uvec), length(uvec), ctx)
+            M = SUNDenseMatrix(Int64(length(uvec)), Int64(length(uvec)), ctx)
             _M = MatrixHandle(M, DenseMatrix())
             if MassLinearSolver === :Dense
                 MLS = SUNLinSol_Dense(utmp, M, ctx)
@@ -815,7 +817,7 @@ function DiffEqBase.__init(prob::DiffEqBase.AbstractODEProblem{uType, tupType, i
             end
         elseif MassLinearSolver in (:Band, :LapackBand)
             nojacobian = false
-            M = SUNBandMatrix(length(uvec), alg.jac_upper, alg.jac_lower, ctx)
+            M = SUNBandMatrix(Int64(length(uvec)), Int64(alg.jac_upper), Int64(alg.jac_lower), ctx)
             _M = MatrixHandle(M, BandMatrix())
             if MassLinearSolver === :Band
                 MLS = SUNLinSol_Band(utmp, M, ctx)
@@ -846,7 +848,8 @@ function DiffEqBase.__init(prob::DiffEqBase.AbstractODEProblem{uType, tupType, i
             _MLS = LinSolHandle(MLS, PTFQMR())
         elseif MassLinearSolver == :KLU
             nnz = length(SparseArrays.nonzeros(prob.f.mass_matrix))
-            M = SUNSparseMatrix(length(uvec), length(uvec), nnz, CSC_MAT, ctx)
+            M = SUNSparseMatrix(
+                Int64(length(uvec)), Int64(length(uvec)), Int64(nnz), CSC_MAT, ctx)
             MLS = SUNLinSol_KLU(utmp, M, ctx)
             _M = MatrixHandle(M, SparseMatrix())
             _MLS = LinSolHandle(MLS, KLU())
@@ -1176,7 +1179,7 @@ function DiffEqBase.__init(
     prec_side = isnothing(alg.prec) ? 0 : 1  # IDA only supports left preconditioning (prec_side = 1)
     if LinearSolver in (:Dense, :LapackDense)
         nojacobian = false
-        A = SUNDenseMatrix(length(u0), length(u0), ctx)
+        A = SUNDenseMatrix(Int64(length(u0)), Int64(length(u0)), ctx)
         _A = MatrixHandle(A, DenseMatrix())
         if LinearSolver === :Dense
             LS = SUNLinSol_Dense(utmp, A, ctx)
@@ -1187,7 +1190,7 @@ function DiffEqBase.__init(
         end
     elseif LinearSolver in (:Band, :LapackBand)
         nojacobian = false
-        A = SUNBandMatrix(length(u0), alg.jac_upper, alg.jac_lower, ctx)
+        A = SUNBandMatrix(Int64(length(u0)), Int64(alg.jac_upper), Int64(alg.jac_lower), ctx)
         _A = MatrixHandle(A, BandMatrix())
         if LinearSolver === :Band
             LS = SUNLinSol_Band(utmp, A, ctx)
@@ -1218,7 +1221,8 @@ function DiffEqBase.__init(
         _LS = LinSolHandle(LS, PTFQMR())
     elseif LinearSolver == :KLU
         nnz = length(SparseArrays.nonzeros(prob.f.jac_prototype))
-        A = SUNSparseMatrix(length(u0), length(u0), nnz, Sundials.CSC_MAT, ctx)
+        A = SUNSparseMatrix(
+            Int64(length(u0)), Int64(length(u0)), Int64(nnz), Sundials.CSC_MAT, ctx)
         LS = SUNLinSol_KLU(utmp, A, ctx)
         _A = MatrixHandle(A, SparseMatrix())
         _LS = LinSolHandle(LS, KLU())
